Added MemPool interface and added a simple version of the mempool.
