+++ src/test/java/com/ibm/darpc/examples/server/DaRPCServer.java
@@ -34,6 +34,7 @@ import org.apache.commons.cli.Options;
 import org.apache.commons.cli.ParseException;
 
 import com.ibm.darpc.DaRPCMemPool;
+import com.ibm.darpc.DaRPCMemPoolImplBuddy;
 import com.ibm.darpc.DaRPCServerEndpoint;
 import com.ibm.darpc.DaRPCServerGroup;
 import com.ibm.darpc.examples.protocol.RdmaRpcRequest;
@@ -41,7 +42,7 @@ import com.ibm.darpc.examples.protocol.RdmaRpcResponse;
 import com.ibm.disni.rdma.*;
 
 public class DaRPCServer {
-	private String ipAddress; 
+	private String ipAddress;
 	private int poolsize = 3;
 	private int recvQueue = 16;
 	private int sendQueue = 16;
@@ -50,7 +51,7 @@ public class DaRPCServer {
 	private boolean polling = false;
 	private int maxinline = 0;
 	private int connections = 16;
-	
+
 	public void run() throws Exception{
 		long[] clusterAffinities = new long[poolsize];
 		for (int i = 0; i < poolsize; i++){
@@ -59,16 +60,16 @@ public class DaRPCServer {
 		}
 		System.out.println("running...server " + ipAddress + ", poolsize " + poolsize + ", maxinline " + maxinline + ", polling " + polling + ", recvQueue " + recvQueue + ", sendQueue " + sendQueue + ", wqSize " + wqSize + ", rpcservice-timeout " + servicetimeout);
 		RdmaRpcService rpcService = new RdmaRpcService(servicetimeout);
-		DaRPCMemPool memPool = new DaRPCMemPool(null, 0, -1, -1, -1);
+		DaRPCMemPool memPool = new DaRPCMemPoolImplBuddy();
 		DaRPCServerGroup<RdmaRpcRequest, RdmaRpcResponse> group = DaRPCServerGroup.createServerGroup(rpcService, memPool, clusterAffinities, -1, maxinline, polling, recvQueue, sendQueue, wqSize, 32);
 		RdmaServerEndpoint<DaRPCServerEndpoint<RdmaRpcRequest, RdmaRpcResponse>> serverEp = group.createServerEndpoint();
 		URI uri = URI.create("rdma://" + ipAddress + ":" + 1919);
 		serverEp.bind(uri);
 		while(true){
 			serverEp.accept();
-		}		
+		}
 	}
-	
+
 	public void launch(String[] args) throws Exception {
 		Option addressOption = Option.builder("a").required().desc("server address").hasArg().build();
 		Option poolsizeOption = Option.builder("p").desc("pool size").hasArg().build();
@@ -132,9 +133,9 @@ public class DaRPCServer {
 		}
 		this.run();
 	}
-	
-	public static void main(String[] args) throws Exception { 
+
+	public static void main(String[] args) throws Exception {
 		DaRPCServer rpcServer = new DaRPCServer();
-		rpcServer.launch(args);		
-	}	
+		rpcServer.launch(args);
+	}
 }
